requirements.txt
python-telegram-bot==21.9
Flask==3.0.3
requests
TgCrypto